Removing prometheus installing
